+++ tests/vectorizer/test_bpemb_vectorizer.py
@@ -4,7 +4,7 @@ from unittest.mock import MagicMock
 
 import numpy as np
 
-from deepparse.data_error import DataError
+from deepparse.errors.data_error import DataError
 from deepparse.embeddings_models.embeddings_model import EmbeddingsModel
 from deepparse.vectorizer import BPEmbVectorizer